backend/tarotMilkTea/tarotdataback/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.1.13
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
pymongo==3.12.1
pyparsing==3.1.2
pytz==2024.1
requests==2.32.3
rsa==4.9
sqlparse==0.2.4
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
